Renaming demo folder to demos to avoid travis collecting demo
